DROOLS-3006 Move tests with equality mode asserts to a separate package
