+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/compiler/integrationtests/AccumulateCepTest.java
@@ -32,14 +32,10 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.kie.api.KieBase;
 import org.kie.api.definition.type.FactType;
-import org.kie.api.io.ResourceType;
 import org.kie.api.runtime.KieSession;
 import org.kie.api.runtime.KieSessionConfiguration;
 import org.kie.api.runtime.conf.ClockTypeOption;
 import org.kie.api.time.SessionPseudoClock;
-import org.kie.internal.builder.KnowledgeBuilder;
-import org.kie.internal.builder.KnowledgeBuilderFactory;
-import org.kie.internal.io.ResourceFactory;
 
 import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
@@ -47,6 +43,33 @@ import static org.junit.Assert.assertFalse;
 
 @RunWith(Parameterized.class)
 public class AccumulateCepTest {
+
+    public static final String TEST_MANY_SLIDING_WINDOWS_DRL = "package com.sample;\n" +
+                "\n" +
+                "global java.util.List list; \n" +
+                "" +
+                "declare Fakt\n" +
+                "  @role( event ) \n" +
+                "  id : int \n" +
+                "end\n" +
+                " \n" +
+                "rule Init \n" +
+                "when \n" +
+                "  $i : Integer() \n" +
+                "then \n" +
+                "  insert( new Fakt( $i ) ); \n" +
+                "end\n" +
+                "" +
+                "rule \"Test\"\n" +
+                "when\n" +
+                "   accumulate ( $d : Fakt( id > 10 ) over window:length(2), $tot1 : count( $d ) ) \n" +
+                "   accumulate ( $d : Fakt( id < 50 ) over window:length(5), $tot2 : count( $d ) ) \n" +
+                "then\n" +
+                "  list.clear();\n " +
+                "  list.add( $tot1.intValue() ); \n" +
+                "  list.add( $tot2.intValue() ); \n" +
+                "end\n" +
+                "\n";
 
     private final KieBaseTestConfiguration kieBaseTestConfiguration;
 
@@ -112,41 +135,17 @@ public class AccumulateCepTest {
 
     @Test
     public void testManySlidingWindows() {
-        final String drl = "package com.sample;\n" +
-                "\n" +
-                "global java.util.List list; \n" +
-                "" +
-                "declare Fakt\n" +
-                "  @role( event ) \n" +
-                "  id : int \n" +
-                "end\n" +
-                " \n" +
-                "rule Init \n" +
-                "when \n" +
-                "  $i : Integer() \n" +
-                "then \n" +
-                "  insert( new Fakt( $i ) ); \n" +
-                "end\n" +
-                "" +
-                "rule \"Test\"\n" +
-                "when\n" +
-                "   accumulate ( $d : Fakt( id > 10 ) over window:length(2), $tot1 : count( $d ) ) \n" +
-                "   accumulate ( $d : Fakt( id < 50 ) over window:length(5), $tot2 : count( $d ) ) \n" +
-                "then\n" +
-                "  list.clear();\n " +
-                "  list.add( $tot1.intValue() ); \n" +
-                "  list.add( $tot2.intValue() ); \n" +
-                "end\n" +
-                "\n";
 
         final KieBase kbase = KieBaseUtil.getKieBaseFromKieModuleFromDrl("accumulate-test", kieBaseTestConfiguration,
-                                                                         drl);
+                                                                         TEST_MANY_SLIDING_WINDOWS_DRL);
         final KieSession ksession = kbase.newKieSession();
         try {
             final List list = new ArrayList();
             ksession.setGlobal("list", list);
 
-            // Intentionally new Integer() here - if using direct value or Integer.valueOf, we get JVM cached values
+            // Intentionally new Integer() here, we need different instances, but some of them equal
+            // - if using direct value or Integer.valueOf, we get JVM cached instances, so we will get the same one for
+            // the same number
             ksession.insert( new Integer( 20 ) );
             ksession.fireAllRules();
             assertEquals(asList(1, 1), list);
@@ -154,44 +153,23 @@ public class AccumulateCepTest {
             ksession.insert(new Integer(20));
             ksession.fireAllRules();
 
-            if (kieBaseTestConfiguration.isIdentity()) {
-                assertEquals(asList(2, 2), list);
+            assertEquals(asList(2, 2), list);
 
-                ksession.insert(new Integer(20));
-                ksession.fireAllRules();
-                assertEquals(asList(2, 3), list);
+            ksession.insert(new Integer(20));
+            ksession.fireAllRules();
+            assertEquals(asList(2, 3), list);
 
-                ksession.insert(new Integer(2));
-                ksession.fireAllRules();
-                assertEquals(asList(2, 4), list);
+            ksession.insert(new Integer(2));
+            ksession.fireAllRules();
+            assertEquals(asList(2, 4), list);
 
-                ksession.insert(new Integer(2));
-                ksession.fireAllRules();
-                assertEquals(asList(2, 5), list);
+            ksession.insert(new Integer(2));
+            ksession.fireAllRules();
+            assertEquals(asList(2, 5), list);
 
-                ksession.insert(new Integer(2));
-                ksession.fireAllRules();
-                assertEquals(asList(2, 5), list);
-            } else {
-                assertEquals(asList(1, 1), list);
-
-                ksession.insert(new Integer(20));
-                ksession.fireAllRules();
-                assertEquals(asList(1, 1), list);
-
-                ksession.insert(new Integer(2));
-                ksession.fireAllRules();
-                assertEquals(asList(1, 2), list);
-
-                ksession.insert(new Integer(2));
-                ksession.fireAllRules();
-                assertEquals(asList(1, 2), list);
-
-                ksession.insert(new Integer(2));
-                ksession.fireAllRules();
-                assertEquals(asList(1, 2), list);
-            }
-
+            ksession.insert(new Integer(2));
+            ksession.fireAllRules();
+            assertEquals(asList(2, 5), list);
         } finally {
             ksession.dispose();
         }
