+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/compiler/integrationtests/BackwardChainingTest.java
@@ -53,252 +53,15 @@ import static org.junit.Assert.fail;
 import static org.kie.api.runtime.rule.Variable.v;
 
 @RunWith(Parameterized.class)
-public class BackwardChainingTest {
-
-    private final KieBaseTestConfiguration kieBaseTestConfiguration;
+public class BackwardChainingTest extends AbstractBackwardChainingTest {
 
     public BackwardChainingTest(final KieBaseTestConfiguration kieBaseTestConfiguration) {
-        this.kieBaseTestConfiguration = kieBaseTestConfiguration;
+        super(kieBaseTestConfiguration);
     }
 
     @Parameterized.Parameters(name = "KieBase type={0}")
     public static Collection<Object[]> getParameters() {
         return TestParametersUtil.getKieBaseCloudConfigurations(true);
-    }
-
-    private String getQueryHeader() {
-        return "package org.drools.compiler.test  \n" +
-                "import " + Person.class.getCanonicalName() + "\n" +
-                "global java.util.List list\n" +
-                "query peeps( String $name, String $likes, int $age ) \n" +
-                "    Person( $name := name, $likes := likes, $age := age; ) \n" +
-                "end\n";
-    }
-
-    @Test(timeout = 10000)
-    public void testQueryPositional() {
-        String drl = getQueryHeader();
-
-        drl += "rule x1\n" +
-                "when\n" +
-                "    String( this == \"go1\" )\n" +
-                //         output        ,output          ,output
-                "    ?peeps($name1 : $name, $likes1 : $likes, $age1 : $age )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        drl += "rule x2\n" +
-                "when\n" +
-                "    String( this == \"go2\" )\n" +
-                //         output, input      ,output
-                "    ?peeps($name1, \"stilton\", $age1; )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        drl += "rule x3\n" +
-                "when\n" +
-                "    String( this == \"go3\" )\n" +
-                "    $name1 : String() from \"darth\";\n " +
-                //         input , input      ,output
-                "    ?peeps($name1, \"stilton\", $age1; )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        drl += "rule x4\n" +
-                "when\n" +
-                "    String( this == \"go4\" )\n" +
-                "    $name1 : String() from \"darth\"\n " +
-                "    $age1 : Integer() from 200;\n " +
-                //         input , input      ,input
-                "    ?peeps($name1, \"stilton\", $age1; )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        testQuery(drl);
-    }
-
-    @Test(timeout = 10000)
-    public void testQueryNamed() {
-        String drl = getQueryHeader();
-
-        drl += "rule x1\n" +
-                "when\n" +
-                "    String( this == \"go1\" )\n" +
-                //         output        ,output          ,output
-                "    ?peeps($name1 : $name, $likes1 : $likes, $age1 : $age )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        drl += "rule x2\n" +
-                "when\n" +
-                "    String( this == \"go2\" )\n" +
-                //         output        ,output                ,output
-                "    ?peeps($name1 : $name, $likes : \"stilton\", $age1 : $age )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        drl += "rule x3\n" +
-                "when\n" +
-                "    String( this == \"go3\" )\n" +
-                "    $name1 : String() from \"darth\";\n " +
-                //         input         ,input                ,output
-                "    ?peeps($name1 : $name, $likes : \"stilton\", $age1 : $age )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        drl += "rule x4\n" +
-                "when\n" +
-                "    String( this == \"go4\" )\n" +
-                "    $name1 : String() from \"darth\";\n " +
-                "    $age1 : Integer() from 200;\n " +
-                //         input         ,input                ,input
-                "    ?peeps($name1 : $name, $likes : \"stilton\", $age1 : $age )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        testQuery(drl);
-    }
-
-    @Test(timeout = 10000)
-    public void testQueryMixed() {
-        String drl = getQueryHeader();
-
-        drl += "rule x1\n" +
-                "when\n" +
-                "    String( this == \"go1\" )\n" +
-                //         output        ,output          ,output
-                "    ?peeps($name1; $likes1 : $likes, $age1 : $age )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        drl += "rule x2\n" +
-                "when\n" +
-                "    String( this == \"go2\" )\n" +
-                //         output        ,output                ,output
-                "    ?peeps($name1, \"stilton\"; $age1 : $age )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        drl += "rule x3\n" +
-                "when\n" +
-                "    String( this == \"go3\" )\n" +
-                "    $name1 : String() from \"darth\";\n " +
-                //         input         ,input                ,output
-                "    ?peeps($name1, \"stilton\"; $age1 : $age )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        drl += "rule x4\n" +
-                "when\n" +
-                "    String( this == \"go4\" )\n" +
-                "    $name1 : String() from \"darth\"\n " +
-                "    $age1 : Integer() from 200;\n " +
-                //         input         ,input                ,input
-                "    ?peeps($name1; $likes : \"stilton\", $age1 : $age )\n" +
-                "then\n" +
-                "   list.add( $name1 + \" : \" + $age1 );\n" +
-                "end \n";
-
-        testQuery(drl);
-    }
-
-    private void testQuery(final String drl) {
-        final KieBase kbase = KieBaseUtil.getKieBaseFromKieModuleFromDrl("backward-chaining-test", kieBaseTestConfiguration, drl);
-        final KieSession ksession = kbase.newKieSession();
-        try {
-            final List<String> list = new ArrayList<>();
-            ksession.setGlobal("list", list);
-
-            final Person p1 = new Person("darth",
-                                         "stilton",
-                                         100);
-            final Person p2 = new Person("darth",
-                                         "stilton",
-                                         200);
-            final Person p3 = new Person("yoda",
-                                         "stilton",
-                                         300);
-            final Person p4 = new Person("luke",
-                                         "brie",
-                                         300);
-            final Person p5 = new Person("bobba",
-                                         "cheddar",
-                                         300);
-
-            ksession.insert(p1);
-            ksession.insert(p2);
-            ksession.insert(p3);
-            ksession.insert(p4);
-            ksession.insert(p5);
-
-            ksession.insert("go1");
-            ksession.fireAllRules();
-            if (kieBaseTestConfiguration.isIdentity()) {
-                assertEquals(5, list.size());
-                assertTrue(list.contains("darth : 100"));
-                assertTrue(list.contains("darth : 200"));
-                assertTrue(list.contains("yoda : 300"));
-                assertTrue(list.contains("luke : 300"));
-                assertTrue(list.contains("bobba : 300"));
-            } else {
-                assertEquals(4, list.size());
-                assertTrue(list.contains("darth : 100"));
-                assertTrue(list.contains("yoda : 300"));
-                assertTrue(list.contains("luke : 300"));
-                assertTrue(list.contains("bobba : 300"));
-            }
-
-            list.clear();
-            ksession.insert("go2");
-            ksession.fireAllRules();
-
-            if (kieBaseTestConfiguration.isIdentity()) {
-                assertEquals(3, list.size());
-                assertTrue(list.contains("darth : 100"));
-                assertTrue(list.contains("darth : 200"));
-                assertTrue(list.contains("yoda : 300"));
-            } else {
-                assertEquals(2, list.size());
-                assertTrue(list.contains("darth : 100"));
-                assertTrue(list.contains("yoda : 300"));
-            }
-
-            list.clear();
-            ksession.insert("go3");
-            ksession.fireAllRules();
-
-            if (kieBaseTestConfiguration.isIdentity()) {
-                assertEquals(2, list.size());
-                assertTrue(list.contains("darth : 100"));
-                assertTrue(list.contains("darth : 200"));
-            } else {
-                assertEquals(1, list.size());
-                assertTrue(list.contains("darth : 100"));
-            }
-
-            list.clear();
-            ksession.insert("go4");
-            ksession.fireAllRules();
-            if (kieBaseTestConfiguration.isIdentity()) {
-                assertEquals(1, list.size());
-                assertTrue(list.contains("darth : 200"));
-            } else {
-                assertEquals(0, list.size());
-            }
-        } finally {
-            ksession.dispose();
-        }
     }
 
     @Test(timeout = 10000)
