+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/compiler/integrationtests/equalitymode/CepEspEqualityModeTest.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2018 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.drools.compiler.integrationtests.equalitymode;
+
+import java.util.Collection;
+
+import org.drools.compiler.integrationtests.AbstractCellTest;
+import org.drools.testcoverage.common.util.EngineTestConfiguration;
+import org.drools.testcoverage.common.util.KieBaseTestConfiguration;
+import org.drools.testcoverage.common.util.TestParametersUtil;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+@RunWith(Parameterized.class)
+public class CepEspEqualityModeTest extends AbstractCellTest {
+
+    public CepEspEqualityModeTest(final KieBaseTestConfiguration kieBaseTestConfiguration) {
+        super(kieBaseTestConfiguration);
+    }
+
+     @Parameterized.Parameters(name = "KieBase type={0}")
+    public static Collection<Object[]> getParameters() {
+        return TestParametersUtil.getKieBaseConfigurations(EngineTestConfiguration.ALPHA_NETWORK_COMPILER_FALSE,
+                                                           EngineTestConfiguration.EQUALITY_MODE,
+                                                           EngineTestConfiguration.STREAM_MODE,
+                                                           EngineTestConfiguration.EXECUTABLE_MODEL_OFF,
+                                                           EngineTestConfiguration.EXECUTABLE_MODEL_FLOW,
+                                                           EngineTestConfiguration.EXECUTABLE_MODEL_PATTERN);
+    }
+}
